# openapi/narip-core.yaml
openapi: 3.1.0
info:
  title: NARIP Core API
  version: 0.1.0
  description: |
    Nexus AI Risk Intelligence Platform — REST surface for seven detection modules,
    unified scoring, integrations, and observability. Live spec is served at `/openapi.json`
    when running the FastAPI application.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/risk/score:
    post:
      summary: Unified enterprise risk (0–100) with SHAP breakdown
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnifiedScoreRequest"
      responses:
        "200":
          description: Unified risk response
  /v1/risk/score/batch:
    post:
      summary: Batch unified scoring (REST alternative to streaming)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items: { $ref: "#/components/schemas/UnifiedScoreRequest" }
      responses:
        "200":
          description: Array of unified risk responses
...
components:
  schemas:
    EmailEvent:
      type: object
      required: [message_id, sender]
      properties:
        message_id: { type: string }
        sender: { type: string }
        subject: { type: string }
        body_text: { type: string }
        urls:
          type: array
          items: { type: string }
    UnifiedScoreRequest:
      type: object
      properties:
        email: { $ref: "#/components/schemas/EmailEvent" }
        transaction: { type: object }
        flows: { type: array }
        otp: { type: object }
        account: { type: object }
        supply: { type: object }